cross evaluation for model trained by srl
